更新了RVM,开始开坑RainC

+++ RVM/SCIInstruction.cpp
@@ -1,6 +1,6 @@
 #include "SCIInstruction.h"
 
-SCIInstruction::SCIInstruction(OptCode optc,int parma1,int parma2,int parma3) : 
+SCIInstruction::SCIInstruction(OptCode optc,size_t parma1,size_t parma2,size_t parma3) : 
 	opt(optc),arg1(parma1),arg2(parma2),arg3(parma3)
 {
